# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Freelance CI API",
        "description": "API REST de la plateforme Freelance CI — connexion freelances, clients et entrepreneurs en Côte d'Ivoire. Paiements via Genius Pay (Mobile Money, Carte), authentification Supabase JWT.",
        "contact": {
            "name": "Freelance CI",
            "email": "contact@freelance-ci.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.6.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://api.freelance-ci.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/boosts/purchase": {
            "post": {
                "tags": [
                    "Boosts"
                ],
                "summary": "Initiate a boost purchase via Genius Pay",
                "operationId": "44d1c1334502a8a1ff0222b575987010",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "target_type",
                                    "duration"
                                ],
                                "properties": {
                                    "target_type": {
                                        "type": "string",
                                        "enum": [
                                            "profile",
                                            "project"
                                        ]
                                    },
                                    "duration": {
                                        "type": "string",
                                        "enum": [
                                            "7_days",
                                            "30_days"
                                        ]
                                    },
                                    "target_id": {
                                        "description": "Project UUID (required if target_type=project)",
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "mobile_money",
                                            "card"
                                        ]
                                    },
                                    "phone": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Boost purchase initiated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction": {
                                                    "$ref": "#/components/schemas/Transaction"
                                                },
                                                "payment_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "reference": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to initiate boost purchase"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/boosts": {
            "get": {
                "tags": [
                    "Boosts"
                ],
                "summary": "List boosts for authenticated user",
                "operationId": "dd4b5ad2ce25e35f494b2eb756bddb19",
                "responses": {
                    "200": {
                        "description": "List of boosts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Boost"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
...
        "/categories/{id}/skills": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "List skills for a category",
                "operationId": "916cd3e29395346dca5f99f6a7b1e194",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of skills",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Skill"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "Skill": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Transaction": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "mission",
                            "subscription",
                            "boost_profile",
                            "boost_project",
                            "badge_verified",
                            "ad",
                            "refund"
                        ]
                    },
                    "amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "payment_channel": {
                        "type": "string",
                        "nullable": true
                    },
                    "payment_operator": {
                        "type": "string",
                        "nullable": true
                    },
                    "operator_status": {
                        "type": "string"
                    },
                    "payment_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "operator_reference": {
                        "type": "string",
                        "nullable": true
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Boost": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "freelance_profile_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "target_type": {
                        "type": "string",
                        "enum": [
                            "profile",
                            "project"
                        ]
                    },
                    "target_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "duration": {
                        "type": "string",
                        "enum": [
                            "7_days",
                            "30_days"
                        ]
                    },
                    "amount_paid": {
                        "type": "number",
                        "format": "float"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "ends_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "description": "JWT Supabase — coller le token reçu après connexion",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "BearerToken": []
        }
    ]
}